requirements.txt
torchdiffeq>=0.1.1
numpy>=1.19.2
scipy>=1.5.2
tqdm>=4.53.0
matplotlib>=3.3.3
pillow>=8.0.1
ninja
imageio==2.9.0
imageio-ffmpeg==0.4.2